+++ test/services/availabilities_service_test.rb
@@ -44,9 +44,10 @@ class AvailabilitiesServiceTest < ActiveSupport::TestCase
     slots = service.trainings([Training.find(2)], @no_subscription, { start: 1.month.ago.beginning_of_day, end: 1.day.ago.end_of_day })
 
     assert_not_empty slots
-    assert_equal Availability.find(20).slots.count, slots.count
-    assert_equal Availability.find(20).start_at, slots.first.start_at
-    assert_equal Availability.find(20).end_at, slots.first.end_at
+    availability = Availability.find(20)
+    assert_equal availability.slots.count, slots.count
+    assert_equal availability.start_at, slots.first.start_at
+    assert_equal availability.end_at, slots.first.end_at
   end
 
   test 'machines availabilities' do
@@ -54,9 +55,10 @@ class AvailabilitiesServiceTest < ActiveSupport::TestCase
     slots = service.machines([Machine.find(1)], @no_subscription, { start: 2.days.from_now.beginning_of_day, end: 4.days.from_now.end_of_day })
 
     assert_not_empty slots
-    assert_equal Availability.find(7).slots.count, slots.count
-    assert_equal Availability.find(7).start_at, slots.first.start_at
-    assert_equal Availability.find(7).end_at, slots.last.end_at
+    availability = Availability.find(7)
+    assert_equal availability.slots.count, slots.count
+    assert_equal availability.start_at, slots.min_by(&:start_at).start_at
+    assert_equal availability.end_at, slots.max_by(&:end_at).end_at
   end
 
   test 'spaces availabilities' do
@@ -64,9 +66,10 @@ class AvailabilitiesServiceTest < ActiveSupport::TestCase
     slots = service.spaces([Space.find(1)], @no_subscription, { start: 2.days.from_now.beginning_of_day, end: 4.days.from_now.end_of_day })
 
     assert_not_empty slots
-    assert_equal Availability.find(18).slots.count, slots.count
-    assert_equal Availability.find(18).start_at, slots.first.start_at
-    assert_equal Availability.find(18).end_at, slots.last.end_at
+    availability = Availability.find(18)
+    assert_equal availability.slots.count, slots.count
+    assert_equal availability.start_at, slots.min_by(&:start_at).start_at
+    assert_equal availability.end_at, slots.max_by(&:end_at).end_at
   end
 
   test 'trainings availabilities' do
@@ -87,8 +90,9 @@ class AvailabilitiesServiceTest < ActiveSupport::TestCase
     slots = service.events([Event.find(4)], @no_subscription, { start: DateTime.current.beginning_of_day, end: 30.days.from_now.end_of_day })
 
     assert_not_empty slots
-    assert_equal Availability.find(17).slots.count, slots.count
-    assert_equal Availability.find(17).start_at, slots.first.start_at
-    assert_equal Availability.find(17).end_at, slots.first.end_at
+    availability = Availability.find(17)
+    assert_equal availability.slots.count, slots.count
+    assert_equal availability.start_at, slots.first.start_at
+    assert_equal availability.end_at, slots.first.end_at
   end
 end