+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improved calendars loading time
 - Refactored and documented the availability-slot-reservation data model
 - Display bookers names to connected users now apply to all resources
+- Fix a bug: unable to reserve if user's subscription plan is disabled
 - Fix a bug: for admins and managers, the current password is not requested before changing their own password
 - Fix a bug: missing translations
 - Fix a bug: unable to book a space's slot with an existing reservation